maandag

+++ src/main/java/com/example/backend/repo/IProductRepository.java
@@ -2,9 +2,9 @@ package com.example.backend.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.backend.Product;
+import com.example.backend.model.Product;
 
-public interface IProductRepository extends JpaRepository<Product, Long>{
-	//was Integer
+public interface IProductRepository extends JpaRepository<Product, Integer>{
+
 
 }
